# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Huginn",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/probe": {
      "post": {
        "summary": "Scrape",
        "description": "Scrape a single URL and return content in requested formats.",
        "operationId": "scrape_v1_probe_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScrapeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScrapeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sweep": {
      "post": {
        "summary": "Start Sweep",
        "description": "Start an async crawl job. Returns job ID for polling, or SSE stream if stream=True.",
        "operationId": "start_sweep_v1_sweep_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrawlRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/jobs": {
      "get": {
        "summary": "List Jobs",
        "description": "List all jobs, optionally filtered by status.",
        "operationId": "list_jobs_v1_jobs_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Action": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/ActionType"
          },
          "selector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selector"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key"
          },
          "direction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction",
            "default": "down"
          },
          "amount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "values": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Values"
          },
          "timeout": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "Action",
        "description": "Browser action to execute before extraction."
      },
      "ActionType": {
        "type": "string",
        "enum": [
          "click",
          "wait",
          "scroll",
          "screenshot",
          "type",
          "press",
          "select",
          "hover",
          "wait_for_selector"
        ],
        "title": "ActionType"
      },
      "CrawlRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "max_depth": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Depth"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "allow_backward_crawling": {
            "type": "boolean",
            "title": "Allow Backward Crawling",
            "default": false
          },
          "allow_external_links": {
            "type": "boolean",
            "title": "Allow External Links",
            "default": false
          },
          "include_paths": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Include Paths"
          },
          "exclude_paths": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exclude Paths"
          },
          "scrape_options": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScrapeOptions"
              },
              {
                "type": "null"
              }
            ]
          },
          "ignore_robots": {
            "type": "boolean",
            "title": "Ignore Robots",
            "description": "Skip robots.txt checks when True",
            "default": false
          },
          "stream": {
            "type": "boolean",
            "title": "Stream",
            "default": false
          },
          "webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Url",
            "description": "URL to POST job completion/failure notifications"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "CrawlRequest",
        "description": "POST /v1/crawl request body."
      },
...
      "DistillOptions": {
        "properties": {
          "prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt"
          },
          "schema_": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema"
          },
          "system_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Prompt"
          }
        },
        "type": "object",
        "title": "DistillOptions",
        "description": "LLM extraction options."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Location": {
        "properties": {
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "languages": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Languages"
          }
        },
        "type": "object",
        "title": "Location",
        "description": "Geographic location for proxy."
      },
...
      "OutputFormat": {
        "type": "string",
        "enum": [
          "markdown",
          "html",
          "raw_html",
          "screenshot",
          "links",
          "metadata",
          "pdf"
        ],
        "title": "OutputFormat"
      },
      "ProxyMode": {
        "type": "string",
        "enum": [
          "basic",
          "stealth"
        ],
        "title": "ProxyMode"
      },
      "ScrapeData": {
        "properties": {
          "markdown": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Markdown"
          },
          "html": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Html"
          },
          "raw_html": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Html"
          },
          "screenshot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Screenshot"
          },
          "links": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Links"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "extract": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extract"
          }
        },
        "type": "object",
        "title": "ScrapeData",
        "description": "Scrape result data."
      },
      "ScrapeOptions": {
        "properties": {
          "formats": {
            "items": {
              "$ref": "#/components/schemas/OutputFormat"
            },
            "type": "array",
            "title": "Formats"
          },
          "only_main_content": {
            "type": "boolean",
            "title": "Only Main Content",
            "default": true
          },
          "include_tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Include Tags"
          },
          "exclude_tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exclude Tags"
          },
          "wait_for": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wait For"
          },
          "timeout": {
            "type": "integer",
            "title": "Timeout",
            "default": 30000
          },
          "headers": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headers"
          },
          "actions": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Action"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actions"
          },
          "max_retries": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 0.0,
            "title": "Max Retries",
            "default": 2
          },
          "scroll": {
            "type": "boolean",
            "title": "Scroll",
            "description": "Auto-scroll page to load dynamic content",
            "default": false
          },
          "render_mode": {
            "type": "string",
            "title": "Render Mode",
            "description": "Rendering mode: auto, full (browser), light (httpx)",
            "default": "auto"
          }
        },
        "type": "object",
        "title": "ScrapeOptions",
        "description": "Options passed to scrape during crawl."
      },
      "ScrapeRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "formats": {
            "items": {
              "$ref": "#/components/schemas/OutputFormat"
            },
            "type": "array",
            "title": "Formats"
          },
          "headers": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headers"
          },
          "wait_for": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wait For"
          },
          "actions": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Action"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actions"
          },
          "extract": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DistillOptions"
              },
              {
                "type": "null"
              }
            ]
          },
          "include_tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Include Tags"
          },
          "exclude_tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exclude Tags"
          },
          "only_main_content": {
            "type": "boolean",
            "title": "Only Main Content",
            "default": true
          },
          "timeout": {
            "type": "integer",
            "title": "Timeout",
            "default": 30000
          },
          "proxy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProxyMode"
              },
              {
                "type": "null"
              }
            ]
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Location"
              },
              {
                "type": "null"
              }
            ]
          },
          "stealth_mode": {
            "type": "boolean",
            "title": "Stealth Mode",
            "default": true
          },
          "max_retries": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 0.0,
            "title": "Max Retries",
            "description": "Max retry attempts on transient errors",
            "default": 2
          },
          "scroll": {
            "type": "boolean",
            "title": "Scroll",
            "description": "Auto-scroll page to load dynamic content before extraction",
            "default": false
          },
          "render_mode": {
            "type": "string",
            "title": "Render Mode",
            "description": "Rendering mode: auto, full (browser), light (httpx)",
            "default": "auto"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "ScrapeRequest",
        "description": "POST /v1/scrape request body."
      },
      "ScrapeResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScrapeData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ScrapeResponse",
        "description": "POST /v1/scrape response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}